Fix process_action() regexs' evaluations (#151)
Before the change, it was possible for the `user_action_regex` conditional to override the `sugg_action_regex` conditional in the event of both of them being true, e.g. when the action is "You #" and both are matched. In such a case, the action text would be used instead of the suggested action text, resulting in a logic error.
This correctly ties the latter conditional to the former by making the `user_speech_regex` conditional `elif` instead of `if`, ensuring that the related conditionals are only evaluated up to the first true one.

Fixes #130 and has been tested with prompt toolkit enabled on a windows terminal CMD interface.